Optimized Logging Feature
-Logging was set to log and clear the string periodically to both reduce
saving to files and reduce the amount of memory the program used.

The larger the memory being stored the more work was caused as it was
counting lines by finding each instance of '\n'.

Instead it now logs based on comparison of string length, as it can be
retrieved efficiently.