Solved an Edge Case Issue with XBR 2x Algorithm
-An x and y offset were being implemented for when the current pixel was
close enough to an edge that one of the surrounding pixels was off
screen would be offset to select the closest one still on screen. Due to
a logic error ALL pixels had been given this offset.

Now there is an x and y offset for the outer pixels as well as the
inner. They will have different offsets depending on whether they're 1
or 2 pixels past the edge. The pixels now apply their offset
appropriately and otherwise do not offset.