Fixed Issue: Sprites Disappearing when Near Screen Top
-Because the (unsigned) byte data type was used to keep track of y,
values 0-255 were available. This meant that when the sprite's top pixel
row was at a negative position, it would be represented as a high value
such as 240. This would mean that it would no longer draw the sprite at
the top of the screen, but at the bottom of the map (not necessarily the
bottom of the window).

This is now solved by using the int data type which allows a wider range
of values including 0-255 and negative values.

-Minor change of subtracting 16 from y on assigning it.

-If statement no longer checks to see if Y is > 8. This is normally
accurate for an emulator but -8 is the equivalent for my emulator. This
also will allow the sprite to appear at the top of the screen as well.
This normally would have stopped it from being drawn, but because the
value was never negative this never stopped it.