Changes to XML Palettes Loading from Main
-Previously the palette setting was not useful because it needed to be 0
to select the single scheme that is loaded without the XML palette file.

Once the XML file was loaded another value was used to select what
scheme it would change to.

Now this value is used to select what palette to change to if the XML
file loads, and if it fails it automatically gets set to 0.

-This code has now been moved to the top of the main function.

-The state of the stream being loaded in is now checked before it is
attempted to be used by loadPalettesFromXML file. This should reduce
potential issues if the loading of the ifstream were to fail.

-Minor spacing issues solved.