Directional Input now ignored during simulatenous Input
-Although the gameboy can do up/down and left/right input at the same
time, it is not possible to do left AND right or up AND down
simultaneously.

On a keyboard this is possible and has to be accounted for to avoid
issues with games when two keys on the same axis are pressed at once.

I have chosen to do this by ignoring input from that axis when both are
pressed, as opposed to preferring one side. (Such as always choosing
right if both are pressed.)