Update to Issue with Palette when XML file is not Loaded
-The code which pads the palette value by one when calling the
returnColor function has been reinstated. Removing it did not solve the
problem but hide the symptoms. It is necessary for some schemes, but was
not relevant to the schemes that had been tested.

-The true issue was the fact that although I had a default color scheme
for when no XML file was loaded, it had never been updated to accomodate
splitting BG0, 0BP0, and 0BP1 into three separate colors. The 4 colors
simply needed to be repeated for 0BP0 and 0BP1 since the code
automatically attempts to index those for sprites.

-Added additonal comment to help explain the purpose of the +1 that is
added to the palette value.