Overload for loadRom that can also Load Save File
-Added an overload for loadRom that contains a bool. If the bool is set
to true the function will check for a .sav file of the same name as the
selected rom and load it if it exists.

This automates the functionality of loading the save file and is fully
optional within the function as well as being an additional function to
the normal loadRom.