Fixed Issue with TIMA incrementing
-To check if the timer is supposed to be enabled the third bit of the
timer register is checked. Masking the register with the value 4 will
isolate the third bit. Previously another step was added to shift the
bit over to become the first bit before comparing it. This was shifting
it by 1 too many and causing the value to always be 0. Because this step
was already redundant (since we're checking if it's greater than 0) this
step has simply been removed completely.

This has fixed games that rely on the timer interrupt, such as Bubble
Ghost.