Improved MBC5 Setting Rom Bank Number
-Previously when the 9th bit of the rom bank number was being set the
whole value would be the operand to a bitwise or operation. If the value
is not 1 or 0, this could be problematic and give bad output.

Instead any value that is not 1 is interpreted as 1, and value is
changed to 1. That way the the bitwise or can only occur with 1 or 0.