Removal of SFML from DLL + Addition of Image & Color Classes
-The next step to creating the DLL core is to remove the SFML
dependency. SFML dependent code has been removed and should be instead
implemented in the frontend.

Some classes are no longer useful now, such as Input and Audio. These
classes will be revamped later on so that they can connect with the
frontend easily. These classes should do none of the actual input
detection or output, but prepare the data in such a way that the
frontend needs very little knowledge of the core to implement them.

-Image class has been created as a replacement for SFML's Image class so
that pixels can be set and retrieved as X, Y coordinates.

-Color class has been created as a replacement for SFML's Color class so
that color settings can be stored as a single object without having to
do additional parsing, such as distribution as an RGBA integer value.