Added LoadRom Function to MainWindow
-Added LoadRom function to open a gb/gbc file. It uses the DLLExports of
the GameBeak dll to transfer the data to the emulator.

It then starts the emulator main function via a new thread. Currently it
sets the emulator to pause so that debugging can be done. Further, this
function should create a thread to start updating the screen. This
screen updating code is currently disabled until errors with it can be
solved.