Reinstituted Input Code
-The input class was temporarily disabled completely due to the fact
that input should be handled by the frontend. While this is still true,
because it has not been implemented yet and testing the core
functionality is currently more important the input class has been
brought back with it no longer checking input keys but automatically
detecting input as no keys pressed. This will prevent issues because the
value will otherwise not get initialized.