DLLExport; Removal of getFirstScreenColor
Removal of no longer needed debug function. Originally it was committed
because the last } was already in the file when this function was added.
(A previous function was remade into this) So if it was deleted from the
file, GitHub still recognized a } being there which would be broken.

Now that this code that used the } has been committed, it can be removed
safely without resulting in any broken code in the file.