GameWindow: Fix for ScrollX Dependent Scenes
The Gameboy draws each line individually and each line can be drawn at
an independent scrollX value to give certain effects. Previously I was
ignoring this and drawing all of the lines at the scrollX value that was
available when it was being drawn.

This was horribly inaccurate and I have partially fixed this by storing
the scrollX value the line had when it was drawn. While this fixes
problems like a scrolling hud in Super Mario Land, it does not fix
problems like redrawing lines since the line is not drawn immediately
like it should.