MainWindow UI: Resized window after removal of QToolBar
The resolution of the gameboy is 160x144. The target resolution by default is 2x (320x288).

The QToolBar has been removed, which reduces some vertical space requirements. However, the menu strip still exists. Further, the grip bar at the bottom has to be factored in so that the full screen will appear in addition to the grip bar.

The new resolution is 320x332.