Gpu: Replaced core, memory, and screen pointers with references
These variables are now references, allowing using the variables  to be a bit simpler. In addition, this change should make the creation of the Gpu object simpler as it will be created alongside the other component objects and pass references to eachother on construction.

Core no longer needs to be passed in since it already has the references it needs.