Removed deprecated/unused class files
Currently the Qt version of the emulator does not have a built in debugger. Eventually equivalents will be added, but likely rewritten entirely. Therefore the related debugger classes are being removed.

Post processing is not currently being used and can be removed. This may be re-implemented later, but there are no current plans.

Audio had never gotten substantial progress and will be removed. It has to be rewritten entirely regardless.

GameWindow has become the Screen class.