Removed Screen widget from MainWindow form
Previously a QGraphicsView was attached to the main window form. It would then be promoted to the Screen subclass.

However, the Screen class will be repurposed to be emulation focused, and a new class will be made to handle displaying frames to a QGraphicsView.

This class will be created in MainWindow and appended during run time.