Disable canvas interactivity
Using the QGraphicsView is perhaps a bit more than what is needed for displaying the frames of the emulator.

No interactivity is required and it is now being disabled to help with performance and reduce work that should not be done.

In the future performance will likely be further improved by using a different medium for rendering the frames.