bgPixels is now std array to conform with other pixel containers
Previously bgPixels was switched to a QVector, however it is used as a standard array and no additional features are required. However, the rest of the pixel containers are still standard arrays and use different syntax. Still, this is acceptable, but it is inconsistent and ideally they would all be treated the same.

That way when changes are made in the future it will be simpler since they do not have to be treated separately.