Program now exits safely on window closing
Previously the program sometimes would not exit safely when emulation was running. This is because the event of the window being closed would be processed during polling  at the start of an emulation loop. This would result in destructors being called, and afterwards it would return to the loop to continue and sigsev due to the object already being deleted.

Instead polling ends at the end of the loop now and the object the loop occurs in will destruct naturally. This way every thing else will be cleaned up in the close event, then when it returns to the event it will see that the run variable has been set to false, exit the function, then destruct, resulting in the program terminating safely and expectedly.