Fixed invalid QColor warnings
Previously the QColor pixel arrays were not initialized properly. The default constructor would be called and the result would be a QColor that is considered invalid.

If this QColor then was rendered, it would generate a warning. This would only occur if the pixels were not overwritten with valid colors before rendering.

Now the arrays are initialized to transparent colors so that there will be no invalid colors.