Added Force DMG and Automatic Selection emulation menu options
Added menu options to the emulator UI that allow the user to control the setting that decides whether a game should be forced into using DMG mode or allow automatic selection based on the game.

If a game has Gameboy Color support it will be selected normally. However, if the Force DMG setting is enabled, this will be overridden and behave as if the game has been put into an original gameboy rather than a gameboy color.