Save files will now be loaded if a .sav file with the same name of the game is found
The functionality to load in save files has existed, however it did not use it by default. Now the default is to check for a .sav file with the same name as the game in the same directory. If one is found it will be loaded in with the game automatically.