Add function to create/overwrite save files from Ram
For games that have persistent save data stored in ram it is now possible to write a .sav file containing the ram data that will be able to be loaded back in.

It accepts a boolean to decide if it should overwrite an existing save file if one is found to exist.

It will attempt to create a save file in the same directly as the rom that has been loaded in with the same name but with the extension ".sav"