Allow loadRom save file paths with periods other than in extension
Previously if a rom path was "file.name.gb" it would attempt to remove the extension of the file then add ".sav" in order to check if a "file.name.sav" file exists.

However, it searched for the first instance of a period, rather than the last. So instead of removing the extension it would remove ".name.gb" as well.

This is fixed. It will now search for the last instance of a period. There is still the possibility of issues for file systems that allow a period after the extension, however it handles likely cases better.