Pressing F1 will now create a save file from Ram data
Pressing F1 will now call the createSaveFile function. If the loaded rom is one that uses the gameboy's ram it can have a file created to store the ram contents.

This will allow the emulator to load it back in for games that use the save ram feature for persistence.

When the file is created it will save a file with the same name as the game loaded but with a  ".sav" extension. If a file already exists with that name it will attempt to overwrite it.