Fixed opengl canvas fading to blue issue
Previously if the camera were to be moved away from the rendering quadrilateral it would start fading to the color of the initial texture frame.

Eventually it would be entirely impossible to see the frame being rendered. This is fine in normal emulation as the camera never needs to be moved, however this also occurred when the canvas window was made smaller. Therefore if scaling the screen down to native GB resolution of 160x144 then the screen would be entirely blue as well.

This has been solved by setting the minification and magnfication filters to linear settings.