Simplified setting opengl min and mag filter settings
Previously the min and mag filter settings were set every frame when the texture was updated to it's new frame. It worked perfectly fine, however it is possible to change these settings once and leave it instead of doing it constantly.

Therefore the settings have been modified in the texture itself now to make it a bit simpler.