Store rom file path for possible reloads and searching for sav files
The rom file path needs to be stored with the rom so that future operations involving the original game file are needed. For example, the path is required for searching for existing .sav files and writing new .sav files. Savestates can fall in this category as well as simply reloading the rom itself.