GameBeakPalette is now dynamic instead of a fixed array
Previously the gameBeakPalette variable was a fixed array. This was restrictive as it was set to a large size so that it would not need to be resized. However, as a QList it can be made larger and smaller as needed and does not have a direct limit of items.