Fixed retrieving palettes.xml path
Previously when retrieving the path of the executable the executable name itself was not removed from the path. This was a problem because we need to check the path for the palettes.xml, and inclusion of the executable name would always result in no file being found.

Now absolutePath is used instead of absoluteFilePath, resulting in the directory as we wanted. Additionally QDir is used to append the palettes file name to the path so that we can check if it exists.

This is useful because absolutePath does not include the ending slash, and which slash to use is platform dependent. QDir handles this for us instead.