Fixed issue with writing palettes.xml file when one does not exist
When there is no palettes.xml file one is written to be used. However, there was a misuse of the QTextStream and QFile leading to the file not having any content written into it.

This is because the << operator was being used on the QFile itself instead of the textstream. This has been fixed and the file is now written successfully.