Moved ColorDialog to its own directory
The Color Dialog was previously located in a Forms folder. However, some additional classes are going to be created that will only be used by the Color Dialog, therefore better organization would be to keep them with the Color Dialog and not with the main emulation classes.

Further, since these new classes will not be forms, they would not make sense to be in the forms folder.