Palettes are now stored in a unified type
Previously when palettes were loaded in they would have their name and colors embedded into different arrays.

Further, the colors were simply arrays of colors, with every 12 colors making up a palette requiring an index to be kept to determine where each begins.

Instead, palettes are now kept in a Palette struct that includes the name and 12 colors.

This way they are associated together and indexing a palette is now much simpler.