ColorDialog current selection is now updated when current row name changes
Previously if one selected a palette in the PaletteListView/QListView, then changed the name, the name of the palette stored in the currentSelection would not get updated.

Therefore it would not match in future checks. This has been fixed. When the dataChanged signal emits a function will be called to check if the currently selected row changed. If it has, it checks is the strings are genuinely different. If so it will update the name stored in currentSelection.