paletteSetting is now a public uint of the GPU class
Previously the paletteSetting started in the core object and was passed as a reference to the gpu class and was kept private.

This is not ideal, as it makes more logical sense to keep it in the GPU with the palette data itself. Further since it was private it could not be accessed by the ColorDialog.

Additionally it has been switched from uchar to uint due to the potential number of palettes exceeding 256.