Created QWidget subclass to act as color selection widget
Added ColorSelectorWidget. This widget will be used to display palette colors when a palette is selected. It will also be used to change a color of the palette via user input. It will give a dialog to select a new RGB color.

For now, it takes the role of the QWidgets that were originally being used for the color previewing.  It internalizes the setColor and getColor functions that were previously done externally.