setPalette now correctly sets the gameBeakPalette colors
Previously this function did correctly set the colors from the current selected palette and overwrite the palette stored in the GPU.

This was not noticed because it had not yet been tested with editing the palette with custom colors. It appeared to be working correctly because the colors on the ColorDialog were identical to what was already stored in the GPU, and when switching the paletteIndex it gave the exact expected results.

This is now fixed and edits to palettes will now show.