ColorSelectorWidget no longer sets color if dialog cancelled
If the color dialog is cancelled the getColor function returns an invalid color. If the dialog is cancelled we don't want to set a color and keep what was originally there. However, without checking if the color is invalid first it will be overwritten on cancel.

This has been fixed.