Fixed not loading all palettes from XML file
Previously there was a bug causing some palettes not to get loaded in. This was primarily due to checking the count of palette names each loop iteration despite the array reducing in size. This means that eventually the array would become smaller while the index would overlap it, meaning that it would stop looping prematurely.

This has been fixed by storing the length before looping and comparing with that number, since it will stay the same, allowing it to loop the correct amount of times.

Additionally there were issues with checking the list's size instead of count when doing earlier loop iterations.