ColorDialog now selects and scrolls to newest palette when one is created
When a new palette was added it was appended to the bottom of the list. If the list was large enough that this occurred off screen then the user might not know anything happened at all.

Now when a new palette is added it will be selected and scrolled to so that they can immediately see it and edit it.