Palette Previews no longer reset to initial palette after reset if previous selection still exists
Previously if a reset was initiated the palette preview would switch to the first palette in the list.

Originally the selection would reset as well. The selection as of a previous change will readjust if the originally selected palette still exists after the reset.

However, now the palette preview will set to the original selection as well.