Removed currentSelection & PaletteSelection since they are no longer needed
Previously the currentSelection variable containing a PaletteSelection struct was used to keep track of the currently selected palette. (Or previously, in cases of deselection)

However, this is no longer useful due to better granular control over the listview and model. Now the listView->currentSelection is reliable for this purpose, so the codebase can be simplified by removing the currentSelection and PaletteSelection.