Fixed Canvas not taking full window space
The default resolution of gamebeak is 320x288. This is exactly twice the size of the native gameboy resolution. When the QHBoxLayout was added the default screen size was scaled to be 300x268. This was pretty similar, however loses the crispness of being the screen size but scaled up.

This has been fixed by disabling margins on the layout.