Update palette name when renamed in Color Dialog list
Previously if the name of a palette in the Color Dialog was changed it would display the new name. However, when the palette was saved to the file it would still have it's previous name.

Now when the list item is renamed the palette corresponding to it will be updated with the new name so that it can be saved properly.

Additionally, if it is renamed to an empty string, it will reset to the previous name.