Restoring selection after color dialog reset now works for new palettes
When a new row is added to the QListView it is not created in the same way that items are created upon initially creating the ListView. Because of this, they will all have their name set in the "text" variable, but not all will have it set as their data.

Because of this, this has been changed to check the text instead of the data since the text will always be there. (And the text and data, assuming it has both, would be identical.)