Fixed out of bounds issues when restoring color dialog selection post-reset
Previously issues were caused by continuing to use a pointer to the previous modelitem after the reset occurs. Now it is deleted and the new one is used.

This stops finding false matches. Also, if the index no longer exists, in the case of a selected palette at the end of the list being removed when resetting, then it will select the first index if one exists.