Pause emulation when loading rom so it is not executed if loading while playing a game
The emulation will now be paused when loading a new rom. This was not the case previously as it was expected that the emulation had not started yet when the rom was loaded. However, if one loads another rom, the previous will be running still.

This is now fixed. This will prevent the execution of the previous rom from writing to the ram while the ram is being re-initialized. This could have led to unexpected behavior.

More work needs to be done to get roms to be able to be loaded after another, as more classes need to be able to reset all of their values.