Removed unused QOpenGlShader objects
Previously the vertexShader and fragmentShader objects were used when creating a shader from source code entered as a string.

However, this is no longer the case. The shaders are now in their own separate source files, which are then used with the shaderProgram object directly. This means that the shader objects are no longer required and can be removed.