Removed blurred look with OpenGL frame texture
Previously the texture being rendered for the emulator frame was slightly blurred. This is because the minification and magnification filters were set to linear instead of nearest. Nearest allows displaying the frame pixel perfect.

Additionally, due to the QLayout attached to the Canvas object an extra window pixel was required to show the full frame vertically correctly. The window as added one pixel to the height.