Corrected window size to accomodate QLayout
The addition of a QLayout to scale the opengl canvas when the window is resized caused the canvas to get squished by 1 pixel on each size. This caused small visual artifiacts due to the screen not being able to fit the full frame in.

To accomodate for this, 2 pixel need to be added to the height and width of the window, allowing enough space to display the whole canvas.