Added Pause and Resume buttons to Menu
Added a Status menu category containing pause and resume actions that can be used.  This will effect the state of emulation. Only one will show at a time. If it is paused, only the resume action will show in the menu, and if it is not paused, only the pause action will show.