Removed unused time variable from fragment shader
Compiling with MingW was resulting in errors due to the time float variable not defining a precision. While this could be addressed simply, this variable was not meant to be included at all. It is not used in the shader and was originally added as an experiment. It is now being removed.