Roms can now be loaded consecutively without restarting
Previously if one loaded a rom to run after having loaded one previously, there was a high chance the new ly loaded rom would not emulate correctly (even if it would if it was the first rom to be loaded).

This is because when loading a new rom not all of the emulation state would be reset to how it was when the emulator was initially started. This means that the gpu and cpu can be in rendering/interrupt states when the previous rom ended and the new rom would still have these states. This would result in extremely unexpected behavior.

To resolve this, each emulation class now has a reset function to reset their state to how it was initally, allowing a blank slate for the next rom to be loaded in with.

Now, when a rom is loaded the state will be reset as expected and the new rom will emulate normally.