No longer rejects save files that are too large
Previously if a save file being loaded in was larger than the save ram it would reject the file. This is not quite necessary. Instead, if the file is larger than save ram, it will simply read in the data it has space for and ignore the rest.

This will allow for .sav files that have trailing data.