Updated loadSaveFile path argument name to reduce confusion
Previously the loadSaveFile function accepted a QString path named simply "path." This could lead to confusion, as one would expect the path to be passed in to be the path directly to the save file.

However, this function takes the path of the rom file and then checks for a .sav file with the same name.

The variable name has been updated to romPath to be clear.