Ensure thatloadSaveFile does not trim more than intended
In order to remove the file extension loadSaveFile removes everything after the '.' however, it would trim everything after the first period. This led to unintentional side effects for files that had multiple in it's name. Now it will only trim after the last instance of a period in the name.