Removed redundancy in save file loading via function
Previously the overload for loadRom that would attempt to load a save file if it existed took the job of finding the save file and loading it in on itself. However, a loadSaveFile function was added for this.

While loadRom IS making use of a loadSaveFile variation, it is making use of the wrong one.
Instead of passing the QByteArray that it loaded in to be directly used by loadSaveFile, it was passing the save file path.

This meant that the function being called would redo all of the work done in the loadRom function.

However, instead of passing the QByteArray, we will simply remove all the loadSaveFile functionality from loadRom and pass in the romFile path to keep their tasks separate and re-useable.