Default File Open path is now OS home path instead of OS dependent /home
Previously the default path for the open file dialog (for selecting a game file) used the Linux specific /home directory.

To make it OS independent it now uses the QDir homePath function in order to attempt to select the home path that the currently used operating system uses.

And if it fails to find the home path it will instead open to the path of the emulator executable.