Include QCloseEvent & QKeyEvent for Qt 5.15.2
Upon updating to 5.15.2 of Qt, it was no longer possible to compile the project with the same source code, due to QCloseEvent being inaccessible when using event->accept(). This has been resolved by including this header.

Also imported QKeyEvent

Also changed a use of NULL to nullptr for C++11 and address warning.